requirements.txt
click==8.1.7
mobase-stubs==2.5.*
pyqt6-stubs @ git+https://github.com/python-qt-tools/PyQt6-stubs
scour==0.38.*
PySide6==6.7.*